pub-sub/dataflow/requirements.txt
apache-beam[gcp]>=2.55.0
google-cloud-pubsub>=2.0.0
